# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pengumuman API",
        "description": "API untuk mengelola pengumuman",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email atau password salah"
                    }
                }
            }
        },
        "/api/register-orangtua": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user orang tua",
                "operationId": "65f264bd13dd992ed0859993aea150a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "nik"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nama Orang Tua"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ortu@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "nik": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NIK tidak ditemukan pada data siswa"
                    },
                    "409": {
                        "description": "Email atau NIK sudah terdaftar"
                    }
                }
            }
        },
        "/api/auth/register-siswa": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register akun siswa baru",
                "operationId": "e5793b5601f8c9e0f3ff84849cbc4c7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "jenis_kelamin",
                                    "no_hp",
                                    "kode_unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ahmad Fauzi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "siswa@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ],
                                        "example": "L"
                                    },
                                    "no_hp": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "kode_unit": {
                                        "type": "string",
                                        "example": "U01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registrasi berhasil"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "no_register": {
                                                            "type": "string",
                                                            "example": "OLU0125X"
                                                        },
                                                        "nama_lengkap": {
                                                            "type": "string",
                                                            "example": "Ahmad Fauzi"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "siswa@email.com"
                                                        },
                                                        "jenis_kelamin": {
                                                            "type": "string",
                                                            "example": "L"
                                                        },
                                                        "no_hp": {
                                                            "type": "string",
                                                            "example": "08123456789"
                                                        },
                                                        "kode_unit": {
                                                            "type": "string",
                                                            "example": "U01"
                                                        },
                                                        "kode_ta": {
                                                            "type": "string",
                                                            "example": "2025"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "token_autentikasi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/getrencanaspp-by-kodebiaya": {
            "get": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Ambil rencana spp berdasarkan kode_biaya",
                "operationId": "9a31f74725ba1a3e906eed2699ec7a43",
                "parameters": [
                    {
                        "name": "kode_biaya",
                        "in": "query",
                        "description": "Kode biaya",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil rencana spp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/gethistoribayar-by-idsiswa": {
            "get": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Ambil histori pembayaran berdasarkan id_siswa",
                "operationId": "6ea685d72e3c80419c6ebaaafa78941a",
                "parameters": [
                    {
                        "name": "id_siswa",
                        "in": "query",
                        "description": "ID Siswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil histori pembayaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/getdetailhistoribayar": {
            "get": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Ambil detail histori pembayaran berdasarkan no_bukti",
                "operationId": "c04f771d3b0d1875d6b697b8ab73c92d",
                "parameters": [
                    {
                        "name": "no_bukti",
                        "in": "query",
                        "description": "Nomor bukti pembayaran",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil detail histori pembayaran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pendaftaranonline/{no_register}/update-foto": {
            "post": {
                "tags": [
                    "Pendaftaran Online"
                ],
                "summary": "Update foto pendaftaran online",
                "description": "Update foto pendaftaran online berdasarkan nomor registrasi.",
                "operationId": "f2bbc3ea7c456f3989d9740b0dc339ca",
                "parameters": [
                    {
                        "name": "no_register",
                        "in": "path",
                        "description": "Nomor registrasi pendaftar yang akan diperbarui.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "foto"
                                ],
                                "properties": {
                                    "foto": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto berhasil diperbarui.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foto berhasil diperbarui."
                                        },
                                        "foto_url": {
                                            "type": "string",
                                            "example": "/storage/foto/pendaftaran/namafile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat memperbarui foto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat memperbarui foto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pendaftaranonline/{id_user}": {
            "get": {
                "tags": [
                    "Pendaftaran Online"
                ],
                "summary": "Ambil data pendaftaran online berdasarkan ID user (autentikasi diperlukan)",
                "description": "Mendapatkan data pendaftaran online berdasarkan ID user.",
                "operationId": "3f96930c6ab909a0197e9a51f28102fd",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pendaftaran online",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/pendaftaranonline/konfirmasi-pembayaran": {
            "post": {
                "tags": [
                    "Pendaftaran Online"
                ],
                "summary": "Konfirmasi pembayaran pendaftaran online",
                "description": "Simpan konfirmasi pembayaran pendaftaran online.",
                "operationId": "579b10a86df2c8d04daf3a6e02b9c1d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "jumlah_pembayaran",
                                    "metode_pembayaran",
                                    "bukti_pembayaran"
                                ],
                                "properties": {
                                    "jumlah_pembayaran": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "metode_pembayaran": {
                                        "type": "string",
                                        "enum": [
                                            "transfer",
                                            "tunai"
                                        ]
                                    },
                                    "bukti_pembayaran": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "keterangan": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konfirmasi pembayaran berhasil disimpan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Konfirmasi pembayaran berhasil disimpan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/public/posts/{slug}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Detail post by slug",
                "operationId": "b955fc8ff877fbe8904c8967899ce4a5",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail data post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detail Data Post Tidak Ditemukan!"
                    }
                }
            }
        },
...
    },
...
}